pyproject.toml
[project]
name = "huggingface-dataset-map"
version = "0.1.0"
description = "Fetch top HuggingFace datasets, embed cards, reduce to 2D, label topics, and visualize"
requires-python = ">=3.10"
dependencies = [
    "huggingface_hub>=0.24",
    "requests>=2.28",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "cohere>=5.0",
    "anthropic>=0.40",
    "numpy>=1.24",
    "umap-learn>=0.5.10",
    "toponymy>=0.4",
    "fast-hdbscan<0.3",
    "datamapplot>=0.5",
    "glasbey>=0.2",
    "joblib>=1.3",
    "tqdm>=4.60",
    "python-dotenv>=1.0",
    "nest-asyncio>=1.5",
    "scikit-learn>=1.2",
    "transformers>=4.30",
    "langcodes>=3.4",
    "language-data>=1.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14",
    "pre-commit>=4.0",
    "pytest>=8.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.lint.per-file-ignores]
"pipeline/03_label_topics.py" = ["E402"]
